Git: always emit wipe-workspace option
In the absence of a `wipe-workspace` declaration, JJB will generate a
WipeWorkspace extension as if `wipe-workspace` was set to `true`.

This differs from Jenkins which does not wipe the workspace unless
instructed to.

Because of the potential confusion, we now always emit `wipe-workspace`
even if it has its default value of `true`.